Update day 15 solution to use rust ranges

+++ day-15/src/circle.rs
@@ -3,7 +3,7 @@ use crate::point::Point;
 #[derive(Clone, Copy)]
 pub struct Circle {
     pub center: Point,
-    pub radius: usize,
+    pub radius: u32,
 }
 
 impl Circle {
